order-service: 用户收货地址相关接口

+++ ruoyi-modules/fy-modules-order/src/main/java/com/cloud/mall/fy/orderservice/entity/UserAddress.java
@@ -16,7 +16,7 @@ public class UserAddress extends BaseEntityForMall {
 
     private String userPhone;
 
-    private Byte defaultFlag;
+    private Integer defaultFlag;
 
     private String provinceName;
